+++ src/wasm.js
@@ -5,9 +5,8 @@ import type {
   SourceMapStringifyOptions,
   IndexedMapping,
 } from "./types";
-
 import path from "path";
-import { generateInlineMap, partialVlqMapToSourceMap } from "./utils";
+import SourceMap from "./SourceMap";
 
 let Module;
 
@@ -40,10 +39,9 @@ function arrayToEmbind(Type, from): any {
   return arr;
 }
 
-export default class SourceMap {
-  sourceMapInstance: any;
-
+export default class WasmSourceMap extends SourceMap {
   constructor() {
+    super();
     this.sourceMapInstance = new Module.SourceMap();
   }
 
@@ -51,19 +49,10 @@ export default class SourceMap {
     sourceName: string,
     sourceContent: string,
     lineOffset: number = 0
-  ): SourceMap {
-    let map = new SourceMap();
+  ): WasmSourceMap {
+    let map = new WasmSourceMap();
     map.addEmptyMap(sourceName, sourceContent, lineOffset);
     return map;
-  }
-
-  addEmptyMap(
-    sourceName: string,
-    sourceContent: string,
-    lineOffset: number = 0
-  ) {
-    this.sourceMapInstance.addEmptyMap(sourceName, sourceContent, lineOffset);
-    return this;
   }
 
   addRawMappings(
@@ -85,87 +74,14 @@ export default class SourceMap {
     return this;
   }
 
-  addBufferMappings(
-    buffer: Buffer,
-    lineOffset: number = 0,
-    columnOffset: number = 0
-  ) {
-    this.sourceMapInstance.addBufferMappings(buffer, lineOffset, columnOffset);
-    return this;
-  }
-
-  // line numbers start at 1 so we have the same api as `source-map` by mozilla
-  addIndexedMappings(
-    mappings: Array<IndexedMapping<number | string>>,
-    lineOffset?: number = 0,
-    columnOffset?: number = 0
-  ) {
-    for (let mapping of mappings) {
-      let hasValidOriginal =
-        mapping.original &&
-        typeof mapping.original.line === "number" &&
-        !isNaN(mapping.original.line) &&
-        typeof mapping.original.column === "number" &&
-        !isNaN(mapping.original.column);
-
-      this.sourceMapInstance.addIndexedMapping(
-        mapping.generated.line + lineOffset - 1,
-        mapping.generated.column + columnOffset,
-        // $FlowFixMe
-        hasValidOriginal ? mapping.original.line - 1 : -1,
-        // $FlowFixMe
-        hasValidOriginal ? mapping.original.column : -1,
-        mapping.source || "",
-        mapping.name || ""
-      );
-    }
-    return this;
-  }
-
-  addName(name: string): number {
-    return this.sourceMapInstance.addName(name);
-  }
-
-  addNames(names: Array<string>): Array<number> {
-    return names.map((n) => this.addName(n));
-  }
-
-  addSource(source: string): number {
-    return this.sourceMapInstance.addSource(source);
-  }
-
-  addSources(sources: Array<string>): Array<number> {
-    return sources.map((s) => this.addSource(s));
-  }
-
-  getSourceIndex(source: string): number {
-    return this.sourceMapInstance.getSourceIndex(source);
-  }
-
-  getSource(index: number): string {
-    return this.sourceMapInstance.getSource(index);
-  }
-
-  getNameIndex(name: string): number {
-    return this.sourceMapInstance.getNameIndex(name);
-  }
-
-  getName(index: number): string {
-    return this.sourceMapInstance.getName(index);
-  }
-
   findClosestMapping(line: number, column: number): ?IndexedMapping<number> {
     let mapping = this.sourceMapInstance.findClosestMapping(line, column);
-    if (mapping.generated.line === -1) return null;
-    else {
+    if (mapping.generated.line === -1) {
+      return null;
+    } else {
       patchMapping(mapping);
       return mapping;
     }
-  }
-
-  extends(buffer: Buffer) {
-    this.sourceMapInstance.extends(buffer);
-    return this;
   }
 
   getMap(): ParsedMap {
@@ -181,20 +97,12 @@ export default class SourceMap {
     };
   }
 
-  toBuffer(): Buffer {
-    return this.sourceMapInstance.toBuffer();
-  }
-
   toVLQ(): VLQMap {
     return {
       mappings: this.sourceMapInstance.getVLQMappings(),
       sources: arrayFromEmbind(this.sourceMapInstance.getSources()),
       names: arrayFromEmbind(this.sourceMapInstance.getNames()),
     };
-  }
-
-  async stringify(options: SourceMapStringifyOptions) {
-    return partialVlqMapToSourceMap(this.toVLQ(), options);
   }
 }
